Implement initial support for Move structs.
This patch implements support for concrete, local struct definitions and usage.
Structs are an involved piece of functionality and consist of 3 main parts:
- Basic concrete structs, defined and used in local scopes (this patch).
- Generic structs (parameterized by types).
- Connection to persistent/global storage.

Generics and global storage are left as future patches. The latter will require
significant discussion and design, as this is where we will need to actually
integrate with the Solana data model.

For local structs, we do the simplest thing possible and treat them as PODs
that are stack-allocated like all other locals. This makes the correspondence
to LLVM IR structs similar to how they would be translated for C++ PODs.

The scheme will need modification or extension to support generics-- at least
if we do not use static instantiations (i.e., when all type parameters are
provided and known to the compiler).

It is not yet clear what modifications will be needed to connect to global
storage. As near as I can tell, Move (at least on Aptos and Sui) always
ser/des global data out/in on access. Under the hood, we can do the same,
where local structs are materialized with the deserialized data, and operated
on as usual. Or any other number of schemes.

One runnable rbpf test is included, where many features have to be correct
for it to possibly run correctly. More runnable and static tests will be
added over time.