Remove native `address_of`.
This is essentially an NFC patch. I added native `address_of` originally
not realizing that `move-stdlib` implements this as Move, not as a native.
So we'll do the same.

Since we do not yet build `move-stdlib` (compiler features still maturing),
signer test cases that want to use `address_of` will need to provide a mock
implementation.