Pull in the transitive closure of called functions.
When declaring all functions, pull in not just the directly called
functions, but the transitive closure of called functions.

Also do not attempt to instantiate generics whose type parameters
themselves are generic (those cannot  be expanded until a function
containing them is instantiated, resolving the type parameters).

This allows to compile std::vector (move-stdlib/sources/vector.move)
for the first time.

Added a runnable test case that includes vector.move and calls a
couple of the non-native methods (singleton and is_empty). This shows
the combination of vector, natives, and generic move functions all
functioning together for this first time!