Implement support for Move struct comparison and Move option.
This patch adds support for comparing Move structures. These are
deep/recursive comparisons so that a field-by-field visitation is
required.

This is currently emitted as LLVM IR, but later we will do it as a fully
move-native runtime call (no such support is yet in the runtime library,
that will be a follow-up task). A runnable rbpf file is included to cover.
When a move-native rt routine is available, the current test will insure
no regressions later.

This patch also includes a runnable test using std::option (from file
move-stdlib/sources/option.move). For the first time, we can now compile
option.move and successfully run rbpf tests using std::option.