Add SBF instruction tracing to the rbpf-tests driver. (#193)
This patch allows the user to turn on SBF instruction tracing/disassembly
for a rbpf case. This is an extremely valuable debugging tool when an rbpf
test crashes in the move-native library-- or perhaps worse-- in core rust
libraries.  This code is copied and integrated from the Solana ledger-tool.

To enable, set environment variable TRACE to a filename where the output
will be directed. Setting TRACE= or TRACE=stdout writes output to stdout:
  TRACE=foo.txt cargo test -p move-mv-llvm-compiler --test rbpf-tests foo-test

Co-authored-by: Dmitri Makarov <dmakarov@users.noreply.github.com>