Fix defects: TargetMachine created too late and DataLayout incorrect.
This patch fixes the two defects in:
https://github.com/solana-labs/move/issues/198.

To summarize: Previously the TargetMachine was created too late
so that the incorrect DataLayout was in effect during translation
(resulting in incorrect alignments, etc). The TargetMachine is
now created before translation and the data layout is set upon
module processing.

Secondly, the target triple and data layout specifiers previously
were never set on the Module when outputting either textual or
bitcode LLVM IR for the module.

This patch also fixes all the "unsafe leakage" from module llvm.
This really is an orthogonal set of changes and should be a
separate patch, however, the custom clippy insisted I do this now.
Wrap up all LLVM references such as LLVMValueRef and so forth
into structs, and reference only those within the translator.

Test results were remastered (eyeball checked for correctness).
The changes are of two kinds (due to the fixes):
1. All LLVM IR expected result files now contain the expected
   "target datalayout" and "target triple" specifiers.
2. Various alignment specifiers on loads and stores increased
   (e.g., 4 to 8). This is as expected and the values were
   previously incorrectly under-aligned.