Generalize and unify LLVM type handling.
Routine llvm_type is currently not general enough to handle generics and
does not properly deal with TypeParameter. In particular, some callers
expect it to special case TypeParameter and treat it as a pointer. This
is not the correct treatment in general, since (by definition) this is a
type filled in by a parameter argument.

During the generics work awhile back, I wrote a wrapper function
llvm_type_with_ty_params that generalized handling of TypeParameter,
and then deferred to the original llvm_type for others. This avoided the
immediate need to audit and update all the users of llvm_type at that time.

This patch unifies the handling into a single routine:
- Audited all call sites of llvm_type. Those that expect special casing
  of TypeParameter now instead explicitly ask for a pointer type and do
  not invoke the type getter. There are only two cases.
- Unified llvm_type and llvm_type_with_ty_params into to_llvm_type.
- Updated remaining callers of llvm_type (those that don't expect special
  casing of TypeParameter, the majority of cases).
- Update all typed-pointer gets to opaque pointer gets (LLVM is now opaque
  pointer mode by default).

Existing tests (especially for generics) already cover this functionality.